Done by recommended fixes
