+++ models/contacts.json
@@ -13,9 +13,9 @@
   },
   {
     "id": "05olLMgyVQdWRwgKfg5J6",
-    "name": "Cyrus Jackson",
-    "email": "nibh@semsempererat.com",
-    "phone": "(501) 472-5218"
+    "name": "Allen ",
+    "email": "nulla.ante@vestibul.co.uk",
+    "phone": "(992) 914-3792"
   },
   {
     "id": "1DEXoP8AuCGYc1YgoQ6hw",
@@ -56,6 +56,12 @@
   {
     "id": "fOHbhpHGSx9fsFIKbD_05",
     "name": "Alex_XAG",
+    "email": "nulla.ante@vestibul.co.uk",
+    "phone": "(992) 914-3792"
+  },
+  {
+    "id": "P9KFsyCvUAxMyFE7DqqTs",
+    "name": "Alex ",
     "email": "nulla.ante@vestibul.co.uk",
     "phone": "(992) 914-3792"
   }
