+++ models/requests.js
@@ -1,0 +1,56 @@
+const fs = require("fs/promises");
+const path = require("path");
+const { nanoid } = require("nanoid");
+
+const contactsPath = path.join(__dirname, "contacts.json");
+
+const getAllContacts = async () => {
+  const data = await fs.readFile(contactsPath);
+
+  return JSON.parse(data);
+};
+
+const getById = async (id) => {
+  const contacts = await getAllContacts();
+  const result = contacts.find((contact) => contact.id === id);
+  return result || null;
+};
+
+const createContact = async (data) => {
+  const contacts = await getAllContacts();
+  const newContact = {
+    id: nanoid(),
+    ...data,
+  };
+  contacts.push(newContact);
+  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
+  return newContact;
+};
+
+const deleteContact = async (id) => {
+  const contacts = await getAllContacts();
+  const index = contacts.findIndex((contact) => contact.id === id);
+  if (index === -1) return null;
+  const [result] = contacts.splice(index, 1);
+  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
+
+  return result;
+};
+
+const changeContact = async (id, body) => {
+  const contacts = await getAllContacts();
+  const index = contacts.findIndex((contact) => contact.id === id);
+  if (index === -1) return null;
+  contacts[index] = { id, ...body };
+  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
+
+  return contacts[index];
+};
+
+module.exports = {
+  getAllContacts,
+  getById,
+  createContact,
+  deleteContact,
+  changeContact,
+};
